requirements.txt
python-dotenv
google-api-python-client
youtube-transcript-api
apify-client